% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcatfish_version.R
\name{rcatfish_version}
\alias{rcatfish_version}
\title{Retrieve the current version Number of the Eschmeyer's Catalog of Fishes Being Used}
\usage{
rcatfish_version()
}
\value{
Character Vector length of one with the Catalog of Fishes version
}
\description{
Returns the current version of Catalog of Fishes being used in the R session
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
\donttest{
#' my.version <- rcatfish_version()
}
\dontshow{\}) # examplesIf}
}
\references{
Fricke, R., Eschmeyer, W.N. & van der Laan, R. (Year Accessed). Eschmeyer's Catalog of Fishes: Genera, Species, References. https://researcharchive.calacademy.org/research/ichthyology/catalog/fishcatmain.asp.
}
\author{
Samuel R. Borstein
}
